+++ core/Inspection.py
@@ -1,4 +1,6 @@
 #!/usr/bin/env python3
+# API Python wrapper for The Vulnerability & Threat Intelligence Feed Service
+# Copyright (C) 2013 - 2020 vFeed, Inc. - https://vfeed.io
 
 import json